backend/requirements.txt
# langchain-openai
pydantic
schedule
python-dotenv
selenium
# openai
# browser-use
pymongo
